Fix typo and add new SearchRatingNotFound exception
Corrected a spelling mistake in the SearchBooruNotFound exception message. Moreover, a new exception called SearchRatingNotFound has been added to handle situations when the specified rating is not found. A 'rating' parameter has been introduced in the 'findit_client' file to improve the search functionality and provide users with the option to filter results based on ratings. This improves the user experience and offers refined search results.