Update version number in __init__.py
The version number in __init__.py file has been updated from '0.3.3.2' to '0.4.0.1'. This update marks the introduction of new features and improvements in the findit_client. Notably, a new exception handling mechanism has been implemented to handle cases when specified rating is not found. This refines the search functionality and enhances user-experience.