Add kwargs to search functions for flexibility
All the image searching methods have been updated to accept keyword arguments (**kwargs), which can be seamlessly passed to the underlying search function. This change improves the functions' flexibility, allowing the user to pass additional arguments when necessary. This is particularly useful when integrating with other APIs or libraries, where the argument list might not be the same. This should not affect the current users as it's backwards compatible.